shift memtrie to separate file
